requirements.txt
autogluon.tabular==1.1.0
streamlit==1.35.0
numpy~=1.26.4
pandas~=2.2.2
requests~=2.32.2
lightgbm==4.3.0
xgboost==2.0.3